requirements.txt
python-docx>=0.8.11
PyPDF2>=3.0.0
pdfplumber>=0.10.0
openai>=1.0.0
anthropic>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0
mcp>=0.9.0
